Initial Slipball implementation
